requirements.txt
decord==0.6.0
einops==0.3.0
ftfy==6.0.3
numpy==1.20.1
pandas==1.1.4
Pillow==8.3.1
regex==2021.4.4
scipy==1.6.2
timm==0.4.5
tqdm==4.60.0
transformers==4.6.0
git+https://github.com/openai/CLIP.git
torch==1.8.1+cu111 -f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.9.1+cu111 -f https://download.pytorch.org/whl/torch_stable.html